requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
audioop-lts==0.2.1
cffi==1.17.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
py-cord==2.6.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
yarl==1.18.3
